better loop of all queries

+++ benchmark-all.sh
@@ -5,24 +5,13 @@ set -e
 DIR="${BASH_SOURCE%/*}"
 if [[ ! -d "${DIR}" ]]; then DIR="${PWD}"; fi
 
-declare -A queries
-for i in q{1..13} q14{a,b} q{15..22} q{23,24}{a,b} q{25..38} q39{a,b} q{40..71} q{73..99} ss_max{,b}
-do
-  queries[$i]=6
-done
+queries=(q{1..13} q14{a,b} q{15..22} q{23,24}{a,b} q{25..38} q39{a,b} q{40..99} ss_max{,b})
+skip=(q4 q11 q13 q14a q14b q16 q17 q20 q23a q23b q24a q24b q25 q29 q50 q64 q67 q72 q95)
 
-queries[q4]=108
-queries[q11]=18
-queries[q64]=12
-queries[q74]=12
-queries[q78]=12
-queries[q80]=12
-queries[q95]=18
-
-for i in q{1..13} q14{a,b} q{15..22} q{23,24}{a,b} q{25..38} q39{a,b} q{40..71} q{73..99} ss_max{,b}
-do
-  p="${queries[$i]}"
-  "${DIR}"/benchmark.sh "${i}" 6 ${p} 2g 2g l
-  "${DIR}"/benchmark.sh "${i}" 6 ${p} 2g 2g p
-  rm ./*.json
+for i in "${queries[@]}"; do
+  if [[ ! "${skip[*]}" =~ ${i} ]]; then
+    "${DIR}"/benchmark.sh "${i}" 4 64 512m 2g l
+    "${DIR}"/benchmark.sh "${i}" 4 64 512m 2g p
+    rm ./*.json
+  fi
 done
